+++ application/views/components/booking_final_step.php
@@ -11,13 +11,13 @@
 <div id="wizard-frame-4" class="wizard-frame" style="display:none;">
     <div class="frame-container">
         <h2 class="frame-title"><?= lang('appointment_confirmation') ?></h2>
-        
+
         <div class="row frame-content m-auto pt-md-4 mb-4">
-            <div id="appointment-details" class="col-12 col-md-6 text-center text-md-start mb-2 mb-md-0">
+            <div id="appointment-details" class="col-12 col-lg-6 text-center text-md-start mb-2 mb-md-0">
                 <!-- JS -->
             </div>
 
-            <div id="customer-details" class="col-12 col-md-6 text-center text-md-end">
+            <div id="customer-details" class="col-12 col-lg-6 text-center text-md-end">
                 <!-- JS -->
             </div>
             <div id="additional-infos" class="col-12 text-start">
@@ -42,8 +42,7 @@
                 </div>
             </div>
         <?php endif; ?>
-        
-        <?php slot('after_captcha'); ?>
+
     </div>
 
     <div class="d-flex fs-6 justify-content-around">
@@ -71,7 +70,6 @@
             </div>
         <?php endif; ?>
 
-        <?php slot('after_select_policies'); ?>
     </div>
     <div>