+++ application/views/components/booking_time_step.php
@@ -15,13 +15,12 @@
         </div>
 
         <div class="row frame-content">
-            <div class="col-12 col-md-6">
+            <div class="col-12 col-lg-6">
                 <div id="select-date"></div>
 
-                <?php slot('after_select_date'); ?>
             </div>
 
-            <div class="col-12 col-md-6">
+            <div class="col-12 col-lg-6">
                 <div id="select-time">
                     <div class="mb-3">
                         <label for="select-timezone" class="form-label">
@@ -33,12 +32,8 @@
                         ]); ?>
                     </div>
 
-                    <?php slot('after_select_timezone'); ?>
-                    
-
                     <div id="available-hours"></div>
 
-                    <?php slot('after_available_hours'); ?>
                 </div>
             </div>
         </div>